+++ etabotsite/etabotapp/TMSlib/JIRA_API.py
@@ -13,6 +13,8 @@ import etabotapp.TMSlib.Atlassian_API as Atlassian_API
 from jira import JIRA
 from typing import Dict
 
+from etabotapp.constants import PROJECTS_AVAILABLE
+
 JIRA_TIMEOUT_FOR_PASSWORD_SECONDS = 10.
 JIRA_TIMEOUT_FOR_OAUTH2_SECONDS = 30.
 JIRA_CLOUD_API = Atlassian_API.ATLASSIAN_CLOUD_BASE + "ex/jira/"
@@ -187,3 +189,21 @@ Errors: "{}"'.format(jira, errors_place))
                 team_members[account_id] = display_name
         logging.debug('Found {} team members: {}.'.format(len(team_members), team_members.keys()))
         return team_members
+
+
+def update_available_projects_for_TMS(tms):
+    jira_wrapper = JIRA_wrapper(
+        tms.endpoint,
+        tms.username,
+        password=tms.password,
+        TMSconfig=tms)
+    projects = jira_wrapper.jira.projects()
+    project_names = [project.name for project in projects]
+    logging.debug('project_names: {}'.format(project_names))
+    logging.debug('TMS: {}'.format(tms))
+    if tms.params is None:
+        tms.params = {}
+    logging.debug('tms.params: {}'.format(tms.params))
+    tms.params[PROJECTS_AVAILABLE] = project_names
+    logging.debug('projects_available: {}'.format(tms.params[PROJECTS_AVAILABLE]))
+    return project_names