+++ etabotsite/etabotapp/TMSlib/TMS.py
@@ -308,6 +308,7 @@ class TMSWrapper(TMS_JIRA):
 
         Arguments:
             tms_config - Django model of TMS.
+            projects - list of Django model projects to pre-populate open_status_values
 
         Todo:
             figure out how to subclass from ProtoTMS to
@@ -348,7 +349,6 @@ projects: {}'.format(tms_config, projects))
                 server_end_point=server,
                 tms_config=tms_config,
                 task_system_schema=task_system_schema)
-            # self.TMS = TMS_JIRA()
         else:
             raise NameError(
                 "TMS_type {} is not supported at this time".format(