+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## v1.3.0 (2021-07-06)
+- added import of only subset of projects
+- added report dates
+- misc fixes
 
 ## v1.1.0 (2021-04-25)
 - Team View